+++ src/predicates/position/in_same_lane_predicate.cpp
@@ -4,7 +4,7 @@
 
 #include "in_same_lane_predicate.h"
 
-bool InSameLanePredicate::booleanEvaluation(int timeStep, const std::shared_ptr<World> &world,
+bool InSameLanePredicate::booleanEvaluation(size_t timeStep, const std::shared_ptr<World> &world,
                                             const std::shared_ptr<Obstacle> &obstacleK,
                                             const std::shared_ptr<Obstacle> &obstacleP) {
     for (const auto &laneP : obstacleP->getOccupiedLanes(world->getRoadNetwork(), timeStep)) {
@@ -18,13 +18,13 @@ bool InSameLanePredicate::booleanEvaluation(int timeStep, const std::shared_ptr<
 
 }
 
-double InSameLanePredicate::robustEvaluation(int timeStep, const std::shared_ptr<World> &world,
+double InSameLanePredicate::robustEvaluation(size_t timeStep, const std::shared_ptr<World> &world,
                                              const std::shared_ptr<Obstacle> &obstacleK,
                                              const std::shared_ptr<Obstacle> &obstacleP) {
     return 0;
 }
 
-Constraint InSameLanePredicate::constraintEvaluation(int timeStep, const std::shared_ptr<World> &world,
+Constraint InSameLanePredicate::constraintEvaluation(size_t timeStep, const std::shared_ptr<World> &world,
                                                      const std::shared_ptr<Obstacle> &obstacleK,
                                                      const std::shared_ptr<Obstacle> &obstacleP) {
     return Constraint();
